Move varied_offset to builders/EUDEMO/tools
